third_party/blink/web_tests/fast/css/fontfaceset-load-expected.txt

Tests load() method of FontFaceSet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS value.length is 2
PASS value[0].unicodeRange is "U+0-10FFFF"
PASS value[1].unicodeRange is "U+41-5A"
PASS value[0].status is "loaded"
PASS value[1].status is "loaded"
PASS value.name is "NetworkError"
PASS value.name is "SyntaxError"
PASS successfullyParsed is true

TEST COMPLETE
